+++ core/web/bridge_types_controller_test.go
@@ -117,7 +117,7 @@ func TestBridgeTypesController_Create_Success(t *testing.T) {
 	bt, err := app.Store.FindBridge(models.MustNewTaskType(btName))
 	assert.NoError(t, err)
 	assert.Equal(t, "randomnumber", bt.Name.String())
-	assert.Equal(t, uint64(10), bt.Confirmations)
+	assert.Equal(t, uint32(10), bt.Confirmations)
 	assert.Equal(t, "https://example.com/randomNumber", bt.URL.String())
 	assert.Equal(t, assets.NewLink(100), bt.MinimumContractPayment)
 	assert.NotEmpty(t, bt.OutgoingToken)
